Crate google_gmail1

This documentation was generated from gmail crate version 0.1.15+20160711, where 20160711 is the exact revision of the gmail:v1 schema built by the mako code generator v0.1.15.

Everything else about the gmail v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

users

drafts create, drafts delete, drafts get, drafts list, drafts send, drafts update, get profile, history list, labels create, labels delete, labels get, labels list, labels patch, labels update, messages attachments get, messages batch delete, messages delete, messages get, messages import, messages insert, messages list, messages modify, messages send, messages trash, messages untrash, settings filters create, settings filters delete, settings filters get, settings filters list, settings forwarding addresses create, settings forwarding addresses delete, settings forwarding addresses get, settings forwarding addresses list, settings get auto forwarding, settings get imap, settings get pop, settings get vacation, settings send as create, settings send as delete, settings send as get, settings send as list, settings send as patch, settings send as update, settings send as verify, settings update auto forwarding, settings update imap, settings update pop, settings update vacation, stop, threads delete, threads get, threads list, threads modify, threads trash, threads untrash and watch

Upload supported by ...

messages import users

drafts create users

drafts send users

messages send users

drafts update users

messages insert users

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.users().messages_untrash(...).doit()
let r = hub.users().messages_get(...).doit()
let r = hub.users().messages_modify(...).doit()
let r = hub.users().messages_import(...).doit()
let r = hub.users().messages_insert(...).doit()
let r = hub.users().messages_send(...).doit()
let r = hub.users().messages_trash(...).doit()
let r = hub.users().drafts_send(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-gmail1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gmail1 as gmail1;
use gmail1::Message;
use gmail1::{Result, Error};
use std::fs;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gmail1::Gmail;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Gmail::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Message::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload(...)`.
// Values shown here are possibly random and not representative !
let result = hub.users().messages_import(req, "userId")
             .process_for_calendar(true)
             .never_mark_spam(true)
             .internal_date_source("diam")
             .deleted(false)
             .upload(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AutoForwarding
Auto-forwarding settings for an account.

BatchDeleteMessagesRequest
There is no detailed description.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Draft
A draft email in the user's mailbox.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Filter
Resource definition for Gmail filters. Filters apply to specific messages instead of an entire email thread.

FilterAction
A set of actions to perform on a message.

FilterCriteria
Message matching criteria.

ForwardingAddress
Settings for a forwarding address.

Gmail
Central instance to access all Gmail related resource activities

History
A record of a change to the user's mailbox. Each history change may affect multiple messages in multiple ways.

HistoryLabelAdded
There is no detailed description.

HistoryLabelRemoved
There is no detailed description.

HistoryMessageAdded
There is no detailed description.

HistoryMessageDeleted
There is no detailed description.

ImapSettings
IMAP settings for an account.

Label
Labels are used to categorize messages and threads within the user's mailbox.

ListDraftsResponse
There is no detailed description.

ListFiltersResponse
Response for the ListFilters method.

ListForwardingAddressesResponse
Response for the ListForwardingAddresses method.

ListHistoryResponse
There is no detailed description.

ListLabelsResponse
There is no detailed description.

ListMessagesResponse
There is no detailed description.

ListSendAsResponse
Response for the ListSendAs method.

ListThreadsResponse
There is no detailed description.

Message
An email message.

MessagePart
A single MIME message part.

MessagePartBody
The body of a single MIME message part.

MessagePartHeader
There is no detailed description.

MethodInfo
Contains information about an API request.

ModifyMessageRequest
There is no detailed description.

ModifyThreadRequest
There is no detailed description.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

PopSettings
POP settings for an account.

Profile
Profile for a Gmail user.

SendAs
Settings associated with a send-as alias, which can be either the primary login address associated with the account or a custom "from" address. Send-as aliases correspond to the "Send Mail As" feature in the web interface. See for more details.

SmtpMsa
Configuration for communication with an SMTP service.

Thread
A collection of messages representing a conversation.

UserDraftCreateCall
Creates a new draft with the DRAFT label.

UserDraftDeleteCall
Immediately and permanently deletes the specified draft. Does not simply trash it.

UserDraftGetCall
Gets the specified draft.

UserDraftListCall
Lists the drafts in the user's mailbox.

UserDraftSendCall
Sends the specified, existing draft to the recipients in the To, Cc, and Bcc headers.

UserDraftUpdateCall
Replaces a draft's content.

UserGetProfileCall
Gets the current user's Gmail profile.

UserHistoryListCall
Lists the history of all changes to the given mailbox. History results are returned in chronological order (increasing historyId).

UserLabelCreateCall
Creates a new label.

UserLabelDeleteCall
Immediately and permanently deletes the specified label and removes it from any messages and threads that it is applied to.

UserLabelGetCall
Gets the specified label.

UserLabelListCall
Lists all labels in the user's mailbox.

UserLabelPatchCall
Updates the specified label. This method supports patch semantics.

UserLabelUpdateCall
Updates the specified label.

UserMessageAttachmentGetCall
Gets the specified message attachment.

UserMessageBatchDeleteCall
Deletes many messages by message ID. Provides no guarantees that messages were not already deleted or even existed at all.

UserMessageDeleteCall
Immediately and permanently deletes the specified message. This operation cannot be undone. Prefer messages.trash instead.

UserMessageGetCall
Gets the specified message.

UserMessageImportCall
Imports a message into only this user's mailbox, with standard email delivery scanning and classification similar to receiving via SMTP. Does not send a message.

UserMessageInsertCall
Directly inserts a message into only this user's mailbox similar to IMAP APPEND, bypassing most scanning and classification. Does not send a message.

UserMessageListCall
Lists the messages in the user's mailbox.

UserMessageModifyCall
Modifies the labels on the specified message.

UserMessageSendCall
Sends the specified message to the recipients in the To, Cc, and Bcc headers.

UserMessageTrashCall
Moves the specified message to the trash.

UserMessageUntrashCall
Removes the specified message from the trash.

UserMethods
A builder providing access to all methods supported on user resources. It is not used directly, but through the Gmail hub.

UserSettingFilterCreateCall
Creates a filter.

UserSettingFilterDeleteCall
Deletes a filter.

UserSettingFilterGetCall
Gets a filter.

UserSettingFilterListCall
Lists the message filters of a Gmail user.

UserSettingForwardingAddresseCreateCall
Creates a forwarding address. If ownership verification is required, a message will be sent to the recipient and the resource's verification status will be set to pending; otherwise, the resource will be created with verification status set to accepted.

UserSettingForwardingAddresseDeleteCall
Deletes the specified forwarding address and revokes any verification that may have been required.

UserSettingForwardingAddresseGetCall
Gets the specified forwarding address.

UserSettingForwardingAddresseListCall
Lists the forwarding addresses for the specified account.

UserSettingGetAutoForwardingCall
Gets the auto-forwarding setting for the specified account.

UserSettingGetImapCall
Gets IMAP settings.

UserSettingGetPopCall
Gets POP settings.

UserSettingGetVacationCall
Gets vacation responder settings.

UserSettingSendACreateCall
Creates a custom "from" send-as alias. If an SMTP MSA is specified, Gmail will attempt to connect to the SMTP service to validate the configuration before creating the alias. If ownership verification is required for the alias, a message will be sent to the email address and the resource's verification status will be set to pending; otherwise, the resource will be created with verification status set to accepted. If a signature is provided, Gmail will sanitize the HTML before saving it with the alias.

UserSettingSendADeleteCall
Deletes the specified send-as alias. Revokes any verification that may have been required for using it.

UserSettingSendAGetCall
Gets the specified send-as alias. Fails with an HTTP 404 error if the specified address is not a member of the collection.

UserSettingSendAListCall
Lists the send-as aliases for the specified account. The result includes the primary send-as address associated with the account as well as any custom "from" aliases.

UserSettingSendAPatchCall
Updates a send-as alias. If a signature is provided, Gmail will sanitize the HTML before saving it with the alias. This method supports patch semantics.

UserSettingSendAUpdateCall
Updates a send-as alias. If a signature is provided, Gmail will sanitize the HTML before saving it with the alias.

UserSettingSendAVerifyCall
Sends a verification email to the specified send-as alias address. The verification status must be pending.

UserSettingUpdateAutoForwardingCall
Updates the auto-forwarding setting for the specified account. A verified forwarding address must be specified when auto-forwarding is enabled.

UserSettingUpdateImapCall
Updates IMAP settings.

UserSettingUpdatePopCall
Updates POP settings.

UserSettingUpdateVacationCall
Updates vacation responder settings.

UserStopCall
Stop receiving push notifications for the given user mailbox.

UserThreadDeleteCall
Immediately and permanently deletes the specified thread. This operation cannot be undone. Prefer threads.trash instead.

UserThreadGetCall
Gets the specified thread.

UserThreadListCall
Lists the threads in the user's mailbox.

UserThreadModifyCall
Modifies the labels applied to the thread. This applies to all messages in the thread.

UserThreadTrashCall
Moves the specified thread to the trash.

UserThreadUntrashCall
Removes the specified thread from the trash.

UserWatchCall
Set up or update a push notification watch on the given user mailbox.

VacationSettings
Vacation auto-reply settings for an account. These settings correspond to the "Vacation responder" feature in the web interface. See for more details.

WatchRequest
Set up or update a new push notification watch on this user's mailbox.

WatchResponse
Push notification watch response.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.