Struct google_gmail1::AutoForwarding

pub struct AutoForwarding {
    pub email_address: Option<String>,
    pub enabled: Option<bool>,
    pub disposition: Option<String>,
}

Auto-forwarding settings for an account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

settings get auto forwarding users (response)

settings update auto forwarding users (request|response)

Fields
email_address: Option<String>

Email address to which all incoming messages are forwarded. This email address must be a verified member of the forwarding addresses.

enabled: Option<bool>

Whether all incoming mail is automatically forwarded to another address.

disposition: Option<String>

The state that a message should be left in after it has been forwarded.