Struct google_gmail1::MessagePart

pub struct MessagePart {
    pub body: Option<MessagePartBody>,
    pub mime_type: Option<String>,
    pub parts: Option<Vec<MessagePart>>,
    pub part_id: Option<String>,
    pub headers: Option<Vec<MessagePartHeader>>,
    pub filename: Option<String>,
}

A single MIME message part.

This type is not used in any activity, and only used as part of another schema.

Fields

body

The message part body for this part, which may be empty for container MIME message parts.

mime_type

The MIME type of the message part.

parts

The child MIME message parts of this part. This only applies to container MIME message parts, for example multipart/*. For non- container MIME message part types, such as text/plain, this field is empty. For more information, see RFC 1521.

part_id

The immutable ID of the message part.

headers

List of headers on this message part. For the top-level message part, representing the entire message payload, it will contain the standard RFC 2822 email headers such as To, From, and Subject.

filename

The filename of the attachment. Only present if this message part represents an attachment.