Struct google_gmail1::SendAs

pub struct SendAs {
    pub smtp_msa: Option<SmtpMsa>,
    pub display_name: Option<String>,
    pub reply_to_address: Option<String>,
    pub signature: Option<String>,
    pub treat_as_alias: Option<bool>,
    pub send_as_email: Option<String>,
    pub is_primary: Option<bool>,
    pub verification_status: Option<String>,
    pub is_default: Option<bool>,
}

Settings associated with a send-as alias, which can be either the primary login address associated with the account or a custom "from" address. Send-as aliases correspond to the "Send Mail As" feature in the web interface. See for more details.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

settings send as create users (request|response)

settings send as update users (request|response)

settings send as patch users (request|response)

settings send as get users (response)

Fields
smtp_msa: Option<SmtpMsa>

An optional SMTP service that will be used as an outbound relay for mail sent using this alias. If this is empty, outbound mail will be sent directly from Gmail's servers to the destination SMTP service. This setting only applies to custom "from" aliases.

display_name: Option<String>

A name that appears in the "From:" header for mail sent using this alias. For custom "from" addresses, when this is empty, Gmail will populate the "From:" header with the name that is used for the primary address associated with the account.

reply_to_address: Option<String>

An optional email address that is included in a "Reply-To:" header for mail sent using this alias. If this is empty, Gmail will not generate a "Reply-To:" header.

signature: Option<String>

An optional HTML signature that is included in messages composed with this alias in the Gmail web UI.

treat_as_alias: Option<bool>

Whether Gmail should treat this address as an alias for the user's primary email address. See for more details. This setting only applies to custom "from" aliases.

send_as_email: Option<String>

The email address that appears in the "From:" header for mail sent using this alias. This is read-only for all operations except create.

is_primary: Option<bool>

Whether this address is the primary address used to login to the account. Every Gmail account has exactly one primary address, and it cannot be deleted from the collection of send-as aliases. This field is read-only.

verification_status: Option<String>

Indicates whether this address has been verified for use as a send-as alias. Read-only. This setting only applies to custom "from" aliases.

is_default: Option<bool>

Whether this address is selected as the default "From:" address in situations such as composing a new message or sending a vacation auto-reply. Every Gmail account has exactly one default send-as address, so the only legal value that clients may write to this field is true. Changing this from false to true for an address will result in this field becoming false for the other previous default address.