Struct google_gmail1::SmtpMsa

pub struct SmtpMsa {
    pub username: Option<String>,
    pub host: Option<String>,
    pub password: Option<String>,
    pub security_mode: Option<String>,
    pub port: Option<i32>,
}

Configuration for communication with an SMTP service.

This type is not used in any activity, and only used as part of another schema.

Fields
username: Option<String>

The username that will be used for authentication with the SMTP service. This is a write-only field that can be specified in requests to create or update SendAs settings; it is never populated in responses.

host: Option<String>

The hostname of the SMTP service. Required.

password: Option<String>

The password that will be used for authentication with the SMTP service. This is a write-only field that can be specified in requests to create or update SendAs settings; it is never populated in responses.

security_mode: Option<String>

The protocol that will be used to secure communication with the SMTP service. Required.

port: Option<i32>

The port of the SMTP service. Required.