Struct google_gmail1::UserHistoryListCall

pub struct UserHistoryListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Lists the history of all changes to the given mailbox. History results are returned in chronological order (increasing historyId).

A builder for the history.list method supported by a user resource. It is not used directly, but through a UserMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.users().history_list("userId")
             .start_history_id("et")
             .page_token("et")
             .max_results(60)
             .label_id("ipsum")
             .doit();

Methods

impl<'a, C, A> UserHistoryListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, ListHistoryResponse)>

Perform the operation you have build so far.

fn user_id(self, new_value: &str) -> UserHistoryListCall<'a, C, A>

The user's email address. The special value me can be used to indicate the authenticated user.

Sets the user id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn start_history_id(self, new_value: &str) -> UserHistoryListCall<'a, C, A>

Required. Returns history records after the specified startHistoryId. The supplied startHistoryId should be obtained from the historyId of a message, thread, or previous list response. History IDs increase chronologically but are not contiguous with random gaps in between valid IDs. Supplying an invalid or out of date startHistoryId typically returns an HTTP 404 error code. A historyId is typically valid for at least a week, but in some rare circumstances may be valid for only a few hours. If you receive an HTTP 404 error response, your application should perform a full sync. If you receive no nextPageToken in the response, there are no updates to retrieve and you can store the returned historyId for a future request.

Sets the start history id query property to the given value.

fn page_token(self, new_value: &str) -> UserHistoryListCall<'a, C, A>

Page token to retrieve a specific page of results in the list.

Sets the page token query property to the given value.

fn max_results(self, new_value: u32) -> UserHistoryListCall<'a, C, A>

The maximum number of history records to return.

Sets the max results query property to the given value.

fn label_id(self, new_value: &str) -> UserHistoryListCall<'a, C, A>

Only return messages with a label matching the ID.

Sets the label id query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> UserHistoryListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> UserHistoryListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> UserHistoryListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Readonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.