Struct google_gmail1::UserMessageImportCall

pub struct UserMessageImportCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Imports a message into only this user's mailbox, with standard email delivery scanning and classification similar to receiving via SMTP. Does not send a message.

A builder for the messages.import method supported by a user resource. It is not used directly, but through a UserMethods instance.

Example

Instantiate a resource method builder

use gmail1::Message;
use std::fs;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Message::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload(...)`.
// Values shown here are possibly random and not representative !
let result = hub.users().messages_import(req, "userId")
             .process_for_calendar(true)
             .never_mark_spam(false)
             .internal_date_source("sadipscing")
             .deleted(true)
             .upload(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());

Methods

impl<'a, C, A> UserMessageImportCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn upload<RS>(self, stream: RS, mime_type: Mime) -> Result<(Response, Message)> where RS: ReadSeek

Upload media all at once. If the upload fails for whichever reason, all progress is lost.

max size: 35MB

multipart: yes

valid mime types: 'message/rfc822'

fn upload_resumable<RS>(self, resumeable_stream: RS, mime_type: Mime) -> Result<(Response, Message)> where RS: ReadSeek

Upload media in a resumable fashion. Even if the upload fails or is interrupted, it can be resumed for a certain amount of time as the server maintains state temporarily.

The delegate will be asked for an upload_url(), and if not provided, will be asked to store an upload URL that was provided by the server, using store_upload_url(...). The upload will be done in chunks, the delegate may specify the chunk_size() and may cancel the operation before each chunk is uploaded, using cancel_chunk_upload(...).

max size: 35MB

multipart: yes

valid mime types: 'message/rfc822'

fn request(self, new_value: Message) -> UserMessageImportCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn user_id(self, new_value: &str) -> UserMessageImportCall<'a, C, A>

The user's email address. The special value me can be used to indicate the authenticated user.

Sets the user id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn process_for_calendar(self, new_value: bool) -> UserMessageImportCall<'a, C, A>

Process calendar invites in the email and add any extracted meetings to the Google Calendar for this user.

Sets the process for calendar query property to the given value.

fn never_mark_spam(self, new_value: bool) -> UserMessageImportCall<'a, C, A>

Ignore the Gmail spam classifier decision and never mark this email as SPAM in the mailbox.

Sets the never mark spam query property to the given value.

fn internal_date_source(self, new_value: &str) -> UserMessageImportCall<'a, C, A>

Source for Gmail's internal date of the message.

Sets the internal date source query property to the given value.

fn deleted(self, new_value: bool) -> UserMessageImportCall<'a, C, A>

Mark the email as permanently deleted (not TRASH) and only visible in Google Apps Vault to a Vault administrator. Only used for Google Apps for Work accounts.

Sets the deleted query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> UserMessageImportCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> UserMessageImportCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> UserMessageImportCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Gmai.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.