Struct google_gmail1::UserMethods

pub struct UserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the Gmail hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gmail1 as gmail1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gmail1::Gmail;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Gmail::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `drafts_create(...)`, `drafts_delete(...)`, `drafts_get(...)`, `drafts_list(...)`, `drafts_send(...)`, `drafts_update(...)`, `get_profile(...)`, `history_list(...)`, `labels_create(...)`, `labels_delete(...)`, `labels_get(...)`, `labels_list(...)`, `labels_patch(...)`, `labels_update(...)`, `messages_attachments_get(...)`, `messages_batch_delete(...)`, `messages_delete(...)`, `messages_get(...)`, `messages_import(...)`, `messages_insert(...)`, `messages_list(...)`, `messages_modify(...)`, `messages_send(...)`, `messages_trash(...)`, `messages_untrash(...)`, `settings_filters_create(...)`, `settings_filters_delete(...)`, `settings_filters_get(...)`, `settings_filters_list(...)`, `settings_forwarding_addresses_create(...)`, `settings_forwarding_addresses_delete(...)`, `settings_forwarding_addresses_get(...)`, `settings_forwarding_addresses_list(...)`, `settings_get_auto_forwarding(...)`, `settings_get_imap(...)`, `settings_get_pop(...)`, `settings_get_vacation(...)`, `settings_send_as_create(...)`, `settings_send_as_delete(...)`, `settings_send_as_get(...)`, `settings_send_as_list(...)`, `settings_send_as_patch(...)`, `settings_send_as_update(...)`, `settings_send_as_verify(...)`, `settings_update_auto_forwarding(...)`, `settings_update_imap(...)`, `settings_update_pop(...)`, `settings_update_vacation(...)`, `stop(...)`, `threads_delete(...)`, `threads_get(...)`, `threads_list(...)`, `threads_modify(...)`, `threads_trash(...)`, `threads_untrash(...)` and `watch(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, A> UserMethods<'a, C, A>

fn settings_update_vacation(&self, request: VacationSettings, user_id: &str) -> UserSettingUpdateVacationCall<'a, C, A>

Create a builder to help you perform the following task:

Updates vacation responder settings.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn labels_create(&self, request: Label, user_id: &str) -> UserLabelCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new label.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_send_as_list(&self, user_id: &str) -> UserSettingSendAListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the send-as aliases for the specified account. The result includes the primary send-as address associated with the account as well as any custom "from" aliases.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn settings_get_imap(&self, user_id: &str) -> UserSettingGetImapCall<'a, C, A>

Create a builder to help you perform the following task:

Gets IMAP settings.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn messages_import(&self, request: Message, user_id: &str) -> UserMessageImportCall<'a, C, A>

Create a builder to help you perform the following task:

Imports a message into only this user's mailbox, with standard email delivery scanning and classification similar to receiving via SMTP. Does not send a message.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_filters_list(&self, user_id: &str) -> UserSettingFilterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the message filters of a Gmail user.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn settings_forwarding_addresses_get(&self, user_id: &str, forwarding_email: &str) -> UserSettingForwardingAddresseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified forwarding address.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

forwardingEmail - The forwarding address to be retrieved.

fn history_list(&self, user_id: &str) -> UserHistoryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the history of all changes to the given mailbox. History results are returned in chronological order (increasing historyId).

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn drafts_create(&self, request: Draft, user_id: &str) -> UserDraftCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new draft with the DRAFT label.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_send_as_get(&self, user_id: &str, send_as_email: &str) -> UserSettingSendAGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified send-as alias. Fails with an HTTP 404 error if the specified address is not a member of the collection.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

sendAsEmail - The send-as alias to be retrieved.

fn labels_delete(&self, user_id: &str, id: &str) -> UserLabelDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified label and removes it from any messages and threads that it is applied to.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to delete.

fn labels_get(&self, user_id: &str, id: &str) -> UserLabelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified label.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to retrieve.

fn labels_list(&self, user_id: &str) -> UserLabelListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all labels in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn drafts_send(&self, request: Draft, user_id: &str) -> UserDraftSendCall<'a, C, A>

Create a builder to help you perform the following task:

Sends the specified, existing draft to the recipients in the To, Cc, and Bcc headers.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_untrash(&self, user_id: &str, id: &str) -> UserMessageUntrashCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the specified message from the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to remove from Trash.

fn settings_get_pop(&self, user_id: &str) -> UserSettingGetPopCall<'a, C, A>

Create a builder to help you perform the following task:

Gets POP settings.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn settings_update_pop(&self, request: PopSettings, user_id: &str) -> UserSettingUpdatePopCall<'a, C, A>

Create a builder to help you perform the following task:

Updates POP settings.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn settings_get_vacation(&self, user_id: &str) -> UserSettingGetVacationCall<'a, C, A>

Create a builder to help you perform the following task:

Gets vacation responder settings.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn messages_batch_delete(&self, request: BatchDeleteMessagesRequest, user_id: &str) -> UserMessageBatchDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes many messages by message ID. Provides no guarantees that messages were not already deleted or even existed at all.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_delete(&self, user_id: &str, id: &str) -> UserMessageDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified message. This operation cannot be undone. Prefer messages.trash instead.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to delete.

fn drafts_get(&self, user_id: &str, id: &str) -> UserDraftGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified draft.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the draft to retrieve.

fn drafts_update(&self, request: Draft, user_id: &str, id: &str) -> UserDraftUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Replaces a draft's content.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the draft to update.

fn labels_update(&self, request: Label, user_id: &str, id: &str) -> UserLabelUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified label.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to update.

fn threads_list(&self, user_id: &str) -> UserThreadListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the threads in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_filters_get(&self, user_id: &str, id: &str) -> UserSettingFilterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a filter.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

id - The ID of the filter to be fetched.

fn labels_patch(&self, request: Label, user_id: &str, id: &str) -> UserLabelPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified label. This method supports patch semantics.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to update.

fn drafts_delete(&self, user_id: &str, id: &str) -> UserDraftDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified draft. Does not simply trash it.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the draft to delete.

fn settings_send_as_create(&self, request: SendAs, user_id: &str) -> UserSettingSendACreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a custom "from" send-as alias. If an SMTP MSA is specified, Gmail will attempt to connect to the SMTP service to validate the configuration before creating the alias. If ownership verification is required for the alias, a message will be sent to the email address and the resource's verification status will be set to pending; otherwise, the resource will be created with verification status set to accepted. If a signature is provided, Gmail will sanitize the HTML before saving it with the alias.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn settings_update_auto_forwarding(&self, request: AutoForwarding, user_id: &str) -> UserSettingUpdateAutoForwardingCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the auto-forwarding setting for the specified account. A verified forwarding address must be specified when auto-forwarding is enabled.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn threads_untrash(&self, user_id: &str, id: &str) -> UserThreadUntrashCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the specified thread from the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to remove from Trash.

fn settings_send_as_patch(&self, request: SendAs, user_id: &str, send_as_email: &str) -> UserSettingSendAPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a send-as alias. If a signature is provided, Gmail will sanitize the HTML before saving it with the alias. This method supports patch semantics.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

sendAsEmail - The send-as alias to be updated.

fn threads_modify(&self, request: ModifyThreadRequest, user_id: &str, id: &str) -> UserThreadModifyCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the labels applied to the thread. This applies to all messages in the thread.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to modify.

fn threads_delete(&self, user_id: &str, id: &str) -> UserThreadDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified thread. This operation cannot be undone. Prefer threads.trash instead.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - ID of the Thread to delete.

fn settings_forwarding_addresses_list(&self, user_id: &str) -> UserSettingForwardingAddresseListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the forwarding addresses for the specified account.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn messages_attachments_get(&self, user_id: &str, message_id: &str, id: &str) -> UserMessageAttachmentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified message attachment.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

messageId - The ID of the message containing the attachment.

id - The ID of the attachment.

fn messages_modify(&self, request: ModifyMessageRequest, user_id: &str, id: &str) -> UserMessageModifyCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the labels on the specified message.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to modify.

fn messages_get(&self, user_id: &str, id: &str) -> UserMessageGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified message.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to retrieve.

fn watch(&self, request: WatchRequest, user_id: &str) -> UserWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Set up or update a push notification watch on the given user mailbox.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_filters_delete(&self, user_id: &str, id: &str) -> UserSettingFilterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a filter.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

id - The ID of the filter to be deleted.

fn drafts_list(&self, user_id: &str) -> UserDraftListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the drafts in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_update_imap(&self, request: ImapSettings, user_id: &str) -> UserSettingUpdateImapCall<'a, C, A>

Create a builder to help you perform the following task:

Updates IMAP settings.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn messages_send(&self, request: Message, user_id: &str) -> UserMessageSendCall<'a, C, A>

Create a builder to help you perform the following task:

Sends the specified message to the recipients in the To, Cc, and Bcc headers.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn threads_trash(&self, user_id: &str, id: &str) -> UserThreadTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Moves the specified thread to the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to Trash.

fn stop(&self, user_id: &str) -> UserStopCall<'a, C, A>

Create a builder to help you perform the following task:

Stop receiving push notifications for the given user mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_forwarding_addresses_create(&self, request: ForwardingAddress, user_id: &str) -> UserSettingForwardingAddresseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a forwarding address. If ownership verification is required, a message will be sent to the recipient and the resource's verification status will be set to pending; otherwise, the resource will be created with verification status set to accepted.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn messages_list(&self, user_id: &str) -> UserMessageListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the messages in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_send_as_update(&self, request: SendAs, user_id: &str, send_as_email: &str) -> UserSettingSendAUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a send-as alias. If a signature is provided, Gmail will sanitize the HTML before saving it with the alias.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

sendAsEmail - The send-as alias to be updated.

fn settings_filters_create(&self, request: Filter, user_id: &str) -> UserSettingFilterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a filter.

Arguments

request - No description provided.

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn messages_trash(&self, user_id: &str, id: &str) -> UserMessageTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Moves the specified message to the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to Trash.

fn get_profile(&self, user_id: &str) -> UserGetProfileCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the current user's Gmail profile.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn threads_get(&self, user_id: &str, id: &str) -> UserThreadGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified thread.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to retrieve.

fn settings_send_as_verify(&self, user_id: &str, send_as_email: &str) -> UserSettingSendAVerifyCall<'a, C, A>

Create a builder to help you perform the following task:

Sends a verification email to the specified send-as alias address. The verification status must be pending.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

sendAsEmail - The send-as alias to be verified.

fn messages_insert(&self, request: Message, user_id: &str) -> UserMessageInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Directly inserts a message into only this user's mailbox similar to IMAP APPEND, bypassing most scanning and classification. Does not send a message.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn settings_get_auto_forwarding(&self, user_id: &str) -> UserSettingGetAutoForwardingCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the auto-forwarding setting for the specified account.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

fn settings_send_as_delete(&self, user_id: &str, send_as_email: &str) -> UserSettingSendADeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified send-as alias. Revokes any verification that may have been required for using it.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

sendAsEmail - The send-as alias to be deleted.

fn settings_forwarding_addresses_delete(&self, user_id: &str, forwarding_email: &str) -> UserSettingForwardingAddresseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified forwarding address and revokes any verification that may have been required.

Arguments

userId - User's email address. The special value "me" can be used to indicate the authenticated user.

forwardingEmail - The forwarding address to be deleted.