Struct google_gmail1::VacationSettings

pub struct VacationSettings {
    pub response_subject: Option<String>,
    pub response_body_plain_text: Option<String>,
    pub restrict_to_contacts: Option<bool>,
    pub enable_auto_reply: Option<bool>,
    pub start_time: Option<String>,
    pub end_time: Option<String>,
    pub response_body_html: Option<String>,
    pub restrict_to_domain: Option<bool>,
}

Vacation auto-reply settings for an account. These settings correspond to the "Vacation responder" feature in the web interface. See for more details.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

settings update vacation users (request|response)

settings get vacation users (response)

Fields
response_subject: Option<String>

Optional text to prepend to the subject line in vacation responses. In order to enable auto-replies, either the response subject or the response body must be nonempty.

response_body_plain_text: Option<String>

Response body in plain text format.

restrict_to_contacts: Option<bool>

Flag that determines whether responses are sent to recipients who are not in the user's list of contacts.

enable_auto_reply: Option<bool>

Flag that controls whether Gmail automatically replies to messages.

start_time: Option<String>

An optional start time for sending auto-replies (epoch ms). When this is specified, Gmail will automatically reply only to messages that it receives after the start time. If both startTime and endTime are specified, startTime must precede endTime.

end_time: Option<String>

An optional end time for sending auto-replies (epoch ms). When this is specified, Gmail will automatically reply only to messages that it receives before the end time. If both startTime and endTime are specified, startTime must precede endTime.

response_body_html: Option<String>

Response body in HTML format. Gmail will sanitize the HTML before storing it.

restrict_to_domain: Option<bool>

Flag that determines whether responses are sent to recipients who are outside of the user's domain. This feature is only available for Google Apps users.