Struct google_gmail1::UserMethods

pub struct UserMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the Gmail hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_gmail1 as gmail1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use gmail1::Gmail;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Gmail::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `drafts_create(...)`, `drafts_delete(...)`, `drafts_get(...)`, `drafts_list(...)`, `drafts_send(...)`, `drafts_update(...)`, `get_profile(...)`, `history_list(...)`, `labels_create(...)`, `labels_delete(...)`, `labels_get(...)`, `labels_list(...)`, `labels_patch(...)`, `labels_update(...)`, `messages_attachments_get(...)`, `messages_delete(...)`, `messages_get(...)`, `messages_import(...)`, `messages_insert(...)`, `messages_list(...)`, `messages_modify(...)`, `messages_send(...)`, `messages_trash(...)`, `messages_untrash(...)`, `stop(...)`, `threads_delete(...)`, `threads_get(...)`, `threads_list(...)`, `threads_modify(...)`, `threads_trash(...)`, `threads_untrash(...)` and `watch(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, A> UserMethods<'a, C, A>

fn stop(&self, user_id: &str) -> UserStopCall<'a, C, A>

Create a builder to help you perform the following task:

Stop receiving push notifications for the given user mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_import(&self, request: Message, user_id: &str) -> UserMessageImportCall<'a, C, A>

Create a builder to help you perform the following task:

Imports a message into only this user's mailbox, with standard email delivery scanning and classification similar to receiving via SMTP. Does not send a message.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn history_list(&self, user_id: &str) -> UserHistoryListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the history of all changes to the given mailbox. History results are returned in chronological order (increasing historyId).

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn drafts_create(&self, request: Draft, user_id: &str) -> UserDraftCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new draft with the DRAFT label.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn labels_create(&self, request: Label, user_id: &str) -> UserLabelCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new label.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn labels_delete(&self, user_id: &str, id: &str) -> UserLabelDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified label and removes it from any messages and threads that it is applied to.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to delete.

fn labels_get(&self, user_id: &str, id: &str) -> UserLabelGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified label.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to retrieve.

fn messages_trash(&self, user_id: &str, id: &str) -> UserMessageTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Moves the specified message to the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to Trash.

fn drafts_send(&self, request: Draft, user_id: &str) -> UserDraftSendCall<'a, C, A>

Create a builder to help you perform the following task:

Sends the specified, existing draft to the recipients in the To, Cc, and Bcc headers.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_untrash(&self, user_id: &str, id: &str) -> UserMessageUntrashCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the specified message from the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to remove from Trash.

fn labels_list(&self, user_id: &str) -> UserLabelListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all labels in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_delete(&self, user_id: &str, id: &str) -> UserMessageDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified message. This operation cannot be undone. Prefer messages.trash instead.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to delete.

fn drafts_update(&self, request: Draft, user_id: &str, id: &str) -> UserDraftUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Replaces a draft's content.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the draft to update.

fn drafts_get(&self, user_id: &str, id: &str) -> UserDraftGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified draft.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the draft to retrieve.

fn labels_update(&self, request: Label, user_id: &str, id: &str) -> UserLabelUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified label.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to update.

fn threads_untrash(&self, user_id: &str, id: &str) -> UserThreadUntrashCall<'a, C, A>

Create a builder to help you perform the following task:

Removes the specified thread from the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to remove from Trash.

fn labels_patch(&self, request: Label, user_id: &str, id: &str) -> UserLabelPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified label. This method supports patch semantics.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the label to update.

fn drafts_delete(&self, user_id: &str, id: &str) -> UserDraftDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified draft. Does not simply trash it.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the draft to delete.

fn threads_list(&self, user_id: &str) -> UserThreadListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the threads in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn threads_modify(&self, request: ModifyThreadRequest, user_id: &str, id: &str) -> UserThreadModifyCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the labels applied to the thread. This applies to all messages in the thread.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to modify.

fn threads_delete(&self, user_id: &str, id: &str) -> UserThreadDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Immediately and permanently deletes the specified thread. This operation cannot be undone. Prefer threads.trash instead.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - ID of the Thread to delete.

fn messages_attachments_get(&self, user_id: &str, message_id: &str, id: &str) -> UserMessageAttachmentGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified message attachment.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

messageId - The ID of the message containing the attachment.

id - The ID of the attachment.

fn messages_modify(&self, request: ModifyMessageRequest, user_id: &str, id: &str) -> UserMessageModifyCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the labels on the specified message.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to modify.

fn threads_trash(&self, user_id: &str, id: &str) -> UserThreadTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Moves the specified thread to the trash.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to Trash.

fn watch(&self, request: WatchRequest, user_id: &str) -> UserWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Set up or update a push notification watch on the given user mailbox.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn drafts_list(&self, user_id: &str) -> UserDraftListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the drafts in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_send(&self, request: Message, user_id: &str) -> UserMessageSendCall<'a, C, A>

Create a builder to help you perform the following task:

Sends the specified message to the recipients in the To, Cc, and Bcc headers.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn messages_get(&self, user_id: &str, id: &str) -> UserMessageGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified message.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the message to retrieve.

fn messages_list(&self, user_id: &str) -> UserMessageListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the messages in the user's mailbox.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn get_profile(&self, user_id: &str) -> UserGetProfileCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the current user's Gmail profile.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

fn threads_get(&self, user_id: &str, id: &str) -> UserThreadGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified thread.

Arguments

userId - The user's email address. The special value me can be used to indicate the authenticated user.

id - The ID of the thread to retrieve.

fn messages_insert(&self, request: Message, user_id: &str) -> UserMessageInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Directly inserts a message into only this user's mailbox similar to IMAP APPEND, bypassing most scanning and classification. Does not send a message.

Arguments

request - No description provided.

userId - The user's email address. The special value me can be used to indicate the authenticated user.